big chunk of typing fixes

+++ src/CynanBot/users/userInterface.py
@@ -84,7 +84,7 @@ class UserInterface(ABC):
         pass
 
     @abstractmethod
-    def getSuperTriviaCheerTriggerAmount(self) -> int | None:
+    def getSuperTriviaCheerTriggerAmount(self) -> float | None:
         pass
 
     @abstractmethod
